% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_head.R
\name{print_head}
\alias{print_head}
\title{fancier version of the head() function}
\usage{
print_head(data)
}
\arguments{
\item{data}{file you want to see}
}
\value{
first 10 rows of a dataset
}
\description{
fancier version of the head() function
}
\examples{
dataset <- data.frame(X1 = c(1.5, 2.8, 3.2, 4.7),X2 = c(5.1, 6.2, 7.3, 8.4),X3 = c(9.9, 10.2, 11.5, 12.8))
print_head(dataset)
}
