% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_from_excel.R
\name{read_data_from_excel}
\alias{read_data_from_excel}
\title{loading in (read) and previewing the data at ones}
\usage{
read_data_from_excel(document, sheet, range)
}
\arguments{
\item{document}{name of the document with type (.xlsx)}

\item{sheet}{which sheet should be loaded in}

\item{range}{the range of the sheet}
}
\value{
loads in data and shows a preview
}
\description{
loading in (read) and previewing the data at ones
}
\examples{
read_data_from_excel("~/dsfb2/dataInspecter/example_files/Annex_1_ESAC-Net_report_2020_downloadable_tables.xlsx", "D1_J01A_AC", "a2:K33")
}
