% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{save_datafile}
\alias{save_datafile}
\title{Save data as .CSV and .rds}
\usage{
save_datafile(data, file_name)
}
\arguments{
\item{data}{a variable with dataset inside}

\item{file_name}{the name the file gets}
}
\value{
saves file as .csv and .rds
}
\description{
Save data as .CSV and .rds
}
\examples{
dataset <- data.frame(X1 = c(1.5, 2.8, 3.2, 4.7), X2 = c(5.1, 6.2, 7.3, 8.4), X3 = c(9.9, 10.2, 11.5, 12.8))
save_datafile(dataset, "test_dataset")
}
