% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_from_excel.R
\name{read_data_from_excel}
\alias{read_data_from_excel}
\title{loading in (read) and previewing the data at ones}
\usage{
read_data_from_excel(document, sheet, range)
}
\arguments{
\item{document}{name of the document with type (.xlsx)}

\item{sheet}{which sheet should be loaded in}

\item{range}{the range of the sheet}
}
\value{
loads in data and shows a preview
}
\description{
loading in (read) and previewing the data at ones
}
\examples{
 test_file <- tempfile(fileext = ".xlsx")
 test_data <- data.frame(X1 = c(1, 2, 3),X2 = c("A", "B", "C"))
 writexl::write_xlsx(test_data, path = test_file)
 read_data_from_excel(test_file, sheet = 1, range = "A1:B4")
}
